adds unit tests for service classes
Unit tests should be run in a lightweight environment, and are supported by most default Gradle configurations (java, groovy, war, etc.).

Since the purpose of this project is to demonstrate the splitting of unit and integration tests, we'd do well having some unit tests in the project.

Additionally, we can specifically test that packages available in runtime (a container in general, or Jetty specifically) environment are not available in a unit test environment.  This should also make it convenient to test the overloading of main, unit, and integration test resources.